Zones overhaul updates (#938)
* remove buildZones function (in favor of using Interface for zoning)
* remove log re: zone creation from reactors.py (buildZones function was removed in prior commit)
* remove reactors.buildZones; move zones.buildManualZones to reactors module
* add unit test coverage, fix Zones.getAllLocations bug, move tests for buildManualZones to test_reactors
* update ThirdCoreHexToFullCoreChanger.convert to adjust zones when growing to full core
