+++ armi/reactor/converters/geometryConverters.py
@@ -1250,14 +1250,17 @@ class ThirdCoreHexToFullCoreChanger(GeometryChanger):
         for a in r.core.getAssemblies():
             # make extras and add them too. since the input is assumed to be 1/3 core.
             otherLocs = grid.getSymmetricEquivalents(a.spatialLocator.indices)
+            thisZone = r.core.zones.findZoneItIsIn(a) if len(r.core.zones) > 0 else None
             angle = 2 * math.pi / (len(otherLocs) + 1)
             count = 1
             for i, j in otherLocs:
                 newAssem = copy.deepcopy(a)
                 newAssem.makeUnique()
                 newAssem.rotate(count * angle)
-                count = count + 1
+                count += 1
                 r.core.add(newAssem, r.core.spatialGrid[i, j, 0])
+                if thisZone:
+                    thisZone.addLoc(newAssem.getLocation())
                 self._newAssembliesAdded.append(newAssem)
 
             if a.getLocation() == "001-001":
